First push

+++ CheckSetup.py
@@ -304,10 +304,10 @@ class AnsibleNetworkingFunctionalityTests(unittest.TestCase):
 
         # Check if any server exists and delete if it does
         existing_server_ids=[item['id'] for item in exec_command_line_command(source_overcloud+'openstack server list --all -f json')['JsonOutput']]
-        self.assertNotEqual(len(existing_server_ids),0,'Failed: no existing servers detected')
-        for id in existing_server_ids:
-            exec_command_line_command(source_overcloud+'openstack server delete '+id)
-        existing_server_ids = [item['id'] for item in exec_command_line_command(source_overcloud+'openstack server list --all -f json')['JsonOutput']]
+        if len(existing_server_ids)==0:
+            for id in existing_server_ids:
+                exec_command_line_command(source_overcloud+'openstack server delete '+id)
+        #existing_server_ids = [item['id'] for item in exec_command_line_command(source_overcloud+'openstack server list --all -f json')['JsonOutput']]
         start_time=time.time()
         to_stop=False
         # Wait till all servers are deleted "
